requirements.txt
streamlit
langchain_openai
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
groq
langchain-groq
sentence_transformers
sse_starlette
